requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
sentence-transformers
llama-index-readers-web 
llama-index-readers-file 
llama-index-vector-stores-chroma 
chromadb